+++ starcitizen/Buttons/LegacyDualAction.cs
@@ -1,0 +1,10 @@
+using BarRaider.SdTools;
+using BarRaider.SdTools.Wrappers;
+
+namespace starcitizen.Buttons
+{
+    [PluginActionId("com.mhwlng.starcitizen.dualaction")]
+    public class LegacyDualAction(SDConnection connection, InitialPayload payload) : DualAction(connection, payload)
+    {
+    }
+}